+++ pcsx2/GS/Renderers/Vulkan/GSDeviceVK.cpp
@@ -4974,7 +4974,7 @@ VkShaderModule GSDeviceVK::GetTFXVertexShader(GSHWDrawConfig::VSSelector sel)
 	AddMacro(ss, "VS_IIP", sel.iip);
 	AddMacro(ss, "VS_POINT_SIZE", sel.point_size);
 	AddMacro(ss, "VS_EXPAND", static_cast<int>(sel.expand));
-	AddMacro(ss, "VS_Z_INTEGER", static_cast<int>(sel.zint));
+	AddMacro(ss, "VS_Z_INTEGER", sel.zint);
 	AddMacro(ss, "VS_PROVOKING_VERTEX_LAST", static_cast<int>(m_features.provoking_vertex_last));
 	ss << m_tfx_source;
